long() -> int()

+++ fastecdsa/test.py
@@ -54,17 +54,17 @@ class TestPrimeFieldCurve(unittest.TestCase):
 
     def test_P384_arith(self):
         S = (
-            long('fba203b81bbd23f2b3be971cc23997e1ae4d89e69cb6f92385dda82768ada415ebab4167459da98e6'
+            int('fba203b81bbd23f2b3be971cc23997e1ae4d89e69cb6f92385dda82768ada415ebab4167459da98e6'
                  '2b1332d1e73cb0e', 16),
-            long('5ffedbaefdeba603e7923e06cdb5d0c65b22301429293376d5c6944e3fa6259f162b4788de6987fd5'
+            int('5ffedbaefdeba603e7923e06cdb5d0c65b22301429293376d5c6944e3fa6259f162b4788de6987fd5'
                  '9aed5e4b5285e45', 16)
         )
-        d = long('a4ebcae5a665983493ab3e626085a24c104311a761b5a8fdac052ed1f111a5c44f76f45659d2d111a'
+        d = int('a4ebcae5a665983493ab3e626085a24c104311a761b5a8fdac052ed1f111a5c44f76f45659d2d111a'
                  '61b5fdd97583480', 16)
         expected = (
-            long('e4f77e7ffeb7f0958910e3a680d677a477191df166160ff7ef6bb5261f791aa7b45e3e653d151b95d'
+            int('e4f77e7ffeb7f0958910e3a680d677a477191df166160ff7ef6bb5261f791aa7b45e3e653d151b95d'
                  'ad3d93ca0290ef2', 16),
-            long('ac7dee41d8c5f4a7d5836960a773cfc1376289d3373f8cf7417b0c6207ac32e913856612fc9ff2e35'
+            int('ac7dee41d8c5f4a7d5836960a773cfc1376289d3373f8cf7417b0c6207ac32e913856612fc9ff2e35'
                  '7eb2ee05cf9667f', 16)
         )
         R = P384.point_mul(S, d)
@@ -73,17 +73,17 @@ class TestPrimeFieldCurve(unittest.TestCase):
 
     def test_P521_arith(self):
         S = (
-            long('000001d5c693f66c08ed03ad0f031f937443458f601fd098d3d0227b4bf62873af50740b0bb84aa15'
+            int('000001d5c693f66c08ed03ad0f031f937443458f601fd098d3d0227b4bf62873af50740b0bb84aa15'
                  '7fc847bcf8dc16a8b2b8bfd8e2d0a7d39af04b089930ef6dad5c1b4', 16),
-            long('00000144b7770963c63a39248865ff36b074151eac33549b224af5c8664c54012b818ed037b2b7c1a'
+            int('00000144b7770963c63a39248865ff36b074151eac33549b224af5c8664c54012b818ed037b2b7c1a'
                  '63ac89ebaa11e07db89fcee5b556e49764ee3fa66ea7ae61ac01823', 16)
         )
-        d = long('000001eb7f81785c9629f136a7e8f8c674957109735554111a2a866fa5a166699419bfa9936c78b62'
+        d = int('000001eb7f81785c9629f136a7e8f8c674957109735554111a2a866fa5a166699419bfa9936c78b62'
                  '653964df0d6da940a695c7294d41b2d6600de6dfcf0edcfc89fdcb1', 16)
         expected = (
-            long('00000091b15d09d0ca0353f8f96b93cdb13497b0a4bb582ae9ebefa35eee61bf7b7d041b8ec34c6c0'
+            int('00000091b15d09d0ca0353f8f96b93cdb13497b0a4bb582ae9ebefa35eee61bf7b7d041b8ec34c6c0'
                  '0c0c0671c4ae063318fb75be87af4fe859608c95f0ab4774f8c95bb', 16),
-            long('00000130f8f8b5e1abb4dd94f6baaf654a2d5810411e77b7423965e0c7fd79ec1ae563c207bd255ee'
+            int('00000130f8f8b5e1abb4dd94f6baaf654a2d5810411e77b7423965e0c7fd79ec1ae563c207bd255ee'
                  '9828eb7a03fed565240d2cc80ddd2cecbb2eb50f0951f75ad87977f', 16)
         )
         R = P521.point_mul(S, d)
